SPU2: Move files from host into main directory
And get rid of the questionable WavFile.
