+++ pcsx2/SPU2/ConfigDebug.cpp
@@ -18,8 +18,7 @@
 #include "pcsx2/Config.h"
 
 #include "SPU2/Global.h"
-#include "SPU2/Host/Dialogs.h"
-#include "SPU2/Host/Config.h"
+#include "SPU2/Config.h"
 #include "common/FileSystem.h"
 #include "common/Path.h"
 #include "common/StringUtil.h"
