+++ pcsx2/pcsx2core.vcxproj
@@ -243,17 +243,18 @@
     <ClCompile Include="Recording\InputRecordingFile.cpp" />
     <ClCompile Include="Recording\PadData.cpp" />
     <ClCompile Include="Recording\Utilities\InputRecordingLogger.cpp" />
+    <ClCompile Include="SPU2\Config.cpp" />
+    <ClCompile Include="SPU2\ConfigDebug.cpp" />
+    <ClCompile Include="SPU2\ConfigSoundTouch.cpp" />
+    <ClCompile Include="SPU2\Debug.cpp" />
+    <ClCompile Include="SPU2\Dma.cpp" />
     <ClCompile Include="SPU2\DplIIdecoder.cpp" />
-    <ClCompile Include="SPU2\debug.cpp" />
-    <ClCompile Include="SPU2\Host\Config.cpp" />
-    <ClCompile Include="SPU2\Host\ConfigDebug.cpp" />
-    <ClCompile Include="SPU2\Host\ConfigSoundTouch.cpp" />
     <ClCompile Include="SPU2\RegLog.cpp" />
     <ClCompile Include="SPU2\SndOut_Cubeb.cpp" />
+    <ClCompile Include="SPU2\SndOut_XAudio2.cpp" />
     <ClCompile Include="SPU2\wavedump_wav.cpp" />
     <ClCompile Include="SPU2\SndOut.cpp" />
     <ClCompile Include="SPU2\Timestretcher.cpp" />
-    <ClCompile Include="SPU2\dma.cpp" />
     <ClCompile Include="SPU2\RegTable.cpp" />
     <ClCompile Include="SPU2\spu2freeze.cpp" />
     <ClCompile Include="SPU2\spu2sys.cpp" />
@@ -330,7 +331,6 @@
     <ClCompile Include="GS\Renderers\Common\GSVertexTraceFMM.cpp" />
     <ClCompile Include="GS\GSXXH.cpp" />
     <ClCompile Include="GS\MultiISA.cpp" />
-    <ClCompile Include="SPU2\Windows\SndOut_XAudio2.cpp" />
     <ClCompile Include="StateWrapper.cpp" />
     <ClCompile Include="USB\deviceproxy.cpp" />
     <ClCompile Include="USB\qemu-usb\bus.cpp" />
@@ -600,14 +600,13 @@
     <ClInclude Include="ShaderCacheVersion.h" />
     <ClInclude Include="SioTypes.h" />
     <ClInclude Include="SPU2\Config.h" />
+    <ClInclude Include="SPU2\Debug.h" />
+    <ClInclude Include="SPU2\Dma.h" />
     <ClInclude Include="SPU2\Global.h" />
-    <ClInclude Include="SPU2\Host\Config.h" />
-    <ClInclude Include="SPU2\Host\Dialogs.h" />
     <ClInclude Include="SPU2\interpolate_table.h" />
     <ClInclude Include="SPU2\SndOut.h" />
     <ClInclude Include="SPU2\spdif.h" />
     <ClInclude Include="SPU2\defs.h" />
-    <ClInclude Include="SPU2\Dma.h" />
     <ClInclude Include="SPU2\regs.h" />
     <ClInclude Include="SPU2\Mixer.h" />
     <ClInclude Include="SPU2\spu2.h" />
